Moved AWS code out and other minor improvements.
* Moved AWS code out of group balancer and into example code
* Removed "unknown" string from UserData...empty string is sufficient
* For uniformity, made RackAffinityGroupBalancer a value receiver
* Minor doc maintenance
